+++ examples/zpp_load/main.cpp
@@ -71,6 +71,11 @@ class Transmitter : public zusi::tx::Base {
 
   // Delay microseconds
   void delayUs(uint32_t us) const final {}
+
+  /// Busy phase
+  ///
+  /// \note Default impl. will block until done
+  virtual void busy() const;
 };
 
 int main() {